+++ crates/aiken-project/src/blueprint/validator.rs
@@ -122,7 +122,7 @@ impl Validator {
                     ),
                 })
             })
-            .collect::<Result<_, _>>()?;
+            .collect::<Result<Vec<_>, _>>()?;
 
         let (datum, redeemer) = if func.name == well_known::VALIDATOR_ELSE {
             (None, None)
@@ -201,6 +201,16 @@ impl Validator {
             title: None,
             schema: Declaration::Inline(Box::new(Schema::Data(Data::Opaque))),
         }));
+
+        definitions
+            .prune_orphan_pairs(
+                parameters
+                    .iter()
+                    .chain(redeemer.as_ref().map(|x| vec![x]).unwrap_or_default())
+                    .chain(datum.as_ref().map(|x| vec![x]).unwrap_or_default())
+                    .collect::<Vec<&Parameter>>(),
+            )
+            .replace_pairs_with_data_lists();
 
         Ok(Validator {
             title: format!("{}.{}.{}", &module.name, &def.name, &func.name,),
@@ -729,6 +739,83 @@ mod tests {
 
             validator recursive_types {
               spend(datum: Option<MyDatum>, redeemer: MyRedeemer<Asset>, output_reference: Data, transaction: Data) {
+                True
+              }
+            }
+            "#
+        );
+    }
+
+    #[test]
+    fn pair_used_after_map() {
+        assert_validator!(
+            r#"
+            pub type MyPair =
+              Pair<Int, Int>
+
+            pub type MyDatum {
+              pairs: List<MyPair>,
+              pair: MyPair,
+            }
+
+            validator placeholder {
+              spend(_datum: Option<MyDatum>, _redeemer: Void, _utxo: Data, _self: Data,) {
+                True
+              }
+            }
+            "#
+        );
+    }
+
+    #[test]
+    fn pair_used_before_map() {
+        assert_validator!(
+            r#"
+            pub type MyPair =
+              Pair<Int, Int>
+
+            pub type MyDatum {
+              pair: MyPair,
+              pairs: List<MyPair>,
+            }
+
+            validator placeholder {
+              spend(_datum: Option<MyDatum>, _redeemer: Void, _utxo: Data, _self: Data,) {
+                True
+              }
+            }
+            "#
+        );
+    }
+
+    #[test]
+    fn map_in_map() {
+        assert_validator!(
+            r#"
+            pub type OuterMap =
+                List<Pair<Int, InnerMap>>
+
+            pub type InnerMap =
+                List<Pair<Int, Bool>>
+
+            validator placeholder {
+              spend(_datum: Option<Void>, _redeemer: OuterMap, _utxo: Data, _self: Data,) {
+                True
+              }
+            }
+            "#
+        );
+    }
+
+    #[test]
+    fn pair_of_lists() {
+        assert_validator!(
+            r#"
+            pub type MyPair =
+              Pair<List<Int>, Bool>
+
+            validator placeholder {
+              spend(_datum: Option<Data>, _redeemer: MyPair, _utxo: Data, _self: Data,) {
                 True
               }
             }